refactor: components types and delete no use code
