+++ apps/frontend/package.json
@@ -15,26 +15,20 @@
     "lint:ci": "next lint --max-warnings=0"
   },
   "dependencies": {
-    "@zk-asset-raffle/sdk": "workspace:*",
     "@zk-asset-raffle/trpc": "workspace:*",
-    "@radix-ui/react-dialog": "^1.1.13",
     "@radix-ui/react-label": "^2.1.6",
     "@radix-ui/react-slot": "^1.2.2",
-    "@radix-ui/react-tabs": "^1.1.11",
     "@radix-ui/react-toast": "^1.2.13",
     "@rainbow-me/rainbowkit": "^2.2.5",
     "@tanstack/react-query": "^5.76.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "date-fns": "^4.1.0",
     "file-saver": "^2.0.5",
     "framer-motion": "^12.23.5",
     "html5-qrcode": "^2.3.8",
     "jszip": "^3.10.1",
     "lucide-react": "^0.511.0",
-    "motion": "^12.23.12",
     "next": "16.1.1",
-    "qrcode": "^1.5.4",
     "qrcode.react": "^4.2.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -52,7 +46,6 @@
     "@types/react-dom": "^19",
     "eslint": "^9",
     "eslint-config-next": "16.1.1",
-    "pino-pretty": "^13.0.0",
     "tailwindcss": "^4",
     "typescript": "^5"
   }
