+++ apps/frontend/src/app/admin/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import AdminDashboard from '@/components/AdminDashboard';
+import AdminDashboard from '@/components/admin/AdminDashboard';
 import { StatsCards } from '@/components/admin/StatsCards';
 
 export default function Page() {
