--- apps/frontend/src/components/RevealKeyDialog.tsx
@@ -1,8 +1,0 @@
-'use client';
-
-import type { RevealKeyComponentProps } from '@/components/reveal/types';
-import { RevealKeyDialogEvm } from '@/components/RevealKeyDialogEvm';
-
-export default function RevealKeyDialog(props: RevealKeyComponentProps) {
-  return <RevealKeyDialogEvm {...props} />;
-}
