+++ apps/frontend/src/components/admin/QRCodeDisplay.tsx
