+++ apps/frontend/src/components/admin/QRCodeSection.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
 import { Alert, AlertDescription } from '@/components/ui/alert';
-import QRCodeDisplay from '@/components/QRCodeDisplay';
+import QRCodeDisplay from '@/components/admin/QRCodeDisplay';
 import { QrCode, RefreshCw, Download, ChevronRight } from 'lucide-react';
 
 export function QRCodeSection({
