+++ apps/frontend/src/components/claim/ClaimForm.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { Card, CardContent } from '@/components/ui/card';
-import QRCodeScanner from '@/components/QRCodeScanner';
+import QRCodeScanner from '@/components/claim/QRCodeScanner';
 
 export default function ClaimForm() {
   const handleScanComplete = () => {
