+++ apps/frontend/src/components/claim/ManualInputPanel.tsx
