+++ apps/frontend/src/components/create/PrizeTierList.tsx
@@ -1,7 +1,7 @@
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
-import { SectionHeader } from "@/components/common/SectionHeader";
+import { SectionHeader } from "@/components/ui/SectionHeader";
 import { Target, PlusCircle, Trash2 } from "lucide-react";
 
 export interface PrizeTier {
