+++ apps/frontend/src/components/layout/ClientProviders.tsx
@@ -4,11 +4,11 @@ import React from "react";
 import dynamic from "next/dynamic";
 import { Header } from "@/components/layout/Header";
 import { Toaster } from "@/components/ui/toaster";
-import { TrpcProvider } from "@/components/TrpcProvider";
+import { TrpcProvider } from "@/components/layout/TrpcProvider";
 
 // Load RainbowKit and Wagmi providers only on the client to avoid SSR IndexedDB access
 const RainbowKitProvider = dynamic(
-  () => import("@/components/RainbowKitProvider").then((m) => m.RainbowKitProvider),
+  () => import("@/components/layout/RainbowKitProvider").then((m) => m.RainbowKitProvider),
   { ssr: false }
 );
