+++ apps/frontend/src/components/layout/brand/ZkAssetRaffleLogo.tsx
