--- apps/frontend/src/components/reveal/types.ts
@@ -1,4 +1,0 @@
-export interface RevealKeyComponentProps {
-  raffleId: string;
-  onRevealSuccess?: (result: { success: boolean; message: string; txHash?: string }) => void;
-}
