--- apps/frontend/src/config/blockchain.ts
@@ -1,3 +1,0 @@
-export type BlockchainTarget = 'evm';
-
-export const BLOCKCHAIN_TARGET: BlockchainTarget = 'evm';
