--- apps/frontend/src/utils/api-service.ts
@@ -1,1 +1,0 @@
-export { apiService } from "@zk-asset-raffle/sdk";
